Rename <Generator> to <JwtGenerator>.
