+++ app/javascript/components/jwt_generator/JwtGenerator.vue
