+++ app/javascript/packs/hello_vue.js
@@ -1,9 +1,9 @@
 import Vue from 'vue/dist/vue.esm'
-import Generator from '../components/jwt_generator/generator.vue'
+import JwtGenerator from '../components/jwt_generator/JwtGenerator.vue'
 
 document.addEventListener('DOMContentLoaded', () => {
   const app = new Vue({
     el: '#hello',
-    components: { Generator }
+    components: { JwtGenerator }
   })
 })
