+++ app/views/layouts/landing.html.erb
@@ -8,7 +8,7 @@
       <%= render partial: 'layouts/partials/notices' %>
 
       <div id='hello'>
-        <Generator></Generator>
+        <jwt-generator></jwt-generator>
       </div>
       <%= yield %>
     </div>
